+++ src/main/java/cl/aacp9/service/PlanService.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface PlanService {
     public List<Plan> findAll();
     public List<Plan> findAllByEstado(Boolean estado);
-    public Optional<Plan> findById(int id);
+    public Optional<Plan> findById(Long id);
     public Plan save(Plan plan);
     public void update(Plan plan);
 }
